refactor(falcon.exceptions): Rename module to falcon.errors
It is (arguably) more Pythonic to talk about "errors" as opposed to
"exceptions". Rename the exceptions module to reflect this.

Closes #342
